+++ src/material.rs
@@ -3,10 +3,10 @@ use crate::prelude::*;
 #[derive(Copy, Clone, Debug, PartialEq)]
 pub struct Material {
     pub color: Color,
-    pub ambient: f32,
-    pub diffuse: f32,
-    pub specular: f32,
-    pub shininess: f32,
+    pub ambient: f64,
+    pub diffuse: f64,
+    pub specular: f64,
+    pub shininess: f64,
 }
 
 impl Default for Material {
@@ -58,7 +58,7 @@ impl Material {
 #[cfg(test)]
 mod test {
     use super::*;
-    use std::f32::consts::FRAC_1_SQRT_2 as S2O2;
+    use std::f64::consts::FRAC_1_SQRT_2 as S2O2;
 
     #[test]
     fn material_default() {
